conf: Move most confvalidators into common package
Previously, some of these validators would not fire correctly.
The package that contributes MUST be imported into frontend, which is not guaranteed. By colocating the GetWarnings func and the validators in one package, this is effectively enforced.

Test plan:

Moved code around, my instance can still start fine and doesn't report new errors.
