+++ cmd/frontend/internal/app/BUILD.bazel
@@ -39,7 +39,6 @@ go_library(
         "//internal/authz",
         "//internal/cloneurls",
         "//internal/conf",
-        "//internal/conf/conftypes",
         "//internal/conf/deploy",
         "//internal/cookie",
         "//internal/database",
@@ -50,7 +49,6 @@ go_library(
         "//internal/httpcli",
         "//internal/otlpenv",
         "//internal/session",
-        "//internal/src-prometheus",
         "//internal/telemetry",
         "//internal/telemetry/teestore",
         "//internal/telemetry/telemetryrecorder",
@@ -94,7 +92,6 @@ go_test(
         "//internal/database/dbmocks",
         "//internal/database/dbtest",
         "//internal/extsvc",
-        "//internal/src-prometheus",
         "//internal/txemail",
         "//internal/types",
         "//lib/errors",
