+++ cmd/frontend/internal/app/debug.go
@@ -21,13 +21,11 @@ import (
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/app/debugproxies"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/app/otlpadapter"
 	"github.com/sourcegraph/sourcegraph/internal/conf"
-	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 	"github.com/sourcegraph/sourcegraph/internal/conf/deploy"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/debugserver"
 	"github.com/sourcegraph/sourcegraph/internal/env"
 	"github.com/sourcegraph/sourcegraph/internal/otlpenv"
-	srcprometheus "github.com/sourcegraph/sourcegraph/internal/src-prometheus"
 	"github.com/sourcegraph/sourcegraph/lib/errors"
 	"github.com/sourcegraph/sourcegraph/schema"
 )
@@ -36,10 +34,6 @@ var (
 	grafanaURLFromEnv = env.Get("GRAFANA_SERVER_URL", "", "URL at which Grafana can be reached")
 	jaegerURLFromEnv  = env.Get("JAEGER_SERVER_URL", "", "URL at which Jaeger UI can be reached")
 )
-
-func init() {
-	conf.ContributeWarning(newPrometheusValidator(srcprometheus.NewClient(srcprometheus.PrometheusURL)))
-}
 
 func addNoK8sClientHandler(r *mux.Router, db database.DB) {
 	noHandler := http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
@@ -290,40 +284,4 @@ func addOpenTelemetryProtocolAdapter(r *mux.Router) {
 
 	// Register adapter endpoints
 	otlpadapter.Register(ctx, logger, protocol, endpoint, r, clientEnabled)
-}
-
-// newPrometheusValidator renders problems with the Prometheus deployment and relevant site configuration
-// as reported by `prom-wrapper` inside the `sourcegraph/prometheus` container if Prometheus is enabled.
-//
-// It also accepts the error from creating `srcprometheus.Client` as an parameter, to validate
-// Prometheus configuration.
-func newPrometheusValidator(prom srcprometheus.Client, promErr error) conf.Validator {
-	return func(c conftypes.SiteConfigQuerier) conf.Problems {
-		// surface new prometheus client error if it was unexpected
-		prometheusUnavailable := errors.Is(promErr, srcprometheus.ErrPrometheusUnavailable)
-		if promErr != nil && !prometheusUnavailable {
-			return conf.NewSiteProblems(fmt.Sprintf("Prometheus (`PROMETHEUS_URL`) might be misconfigured: %v", promErr))
-		}
-
-		// no need to validate prometheus config if no `observability.*` settings are configured
-		observabilityNotConfigured := len(c.SiteConfig().ObservabilityAlerts) == 0 && len(c.SiteConfig().ObservabilitySilenceAlerts) == 0
-		if observabilityNotConfigured {
-			// no observability configuration, no checks to make
-			return nil
-		} else if prometheusUnavailable {
-			// no prometheus, but observability is configured
-			return conf.NewSiteProblems("`observability.alerts` or `observability.silenceAlerts` are configured, but Prometheus is not available")
-		}
-
-		// use a short timeout to avoid having this block problems from loading
-		ctx, cancel := context.WithTimeout(context.Background(), 500*time.Millisecond)
-		defer cancel()
-
-		// get reported problems
-		status, err := prom.GetConfigStatus(ctx)
-		if err != nil {
-			return conf.NewSiteProblems(fmt.Sprintf("`observability`: failed to fetch alerting configuration status: %v", err))
-		}
-		return status.Problems
-	}
 }
