+++ cmd/frontend/internal/authz/init.go
@@ -12,7 +12,6 @@ import (
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/graphqlbackend"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/authz/resolvers"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/authz/webhooks"
-	"github.com/sourcegraph/sourcegraph/internal/actor"
 	"github.com/sourcegraph/sourcegraph/internal/authz"
 	"github.com/sourcegraph/sourcegraph/internal/authz/providers"
 	srp "github.com/sourcegraph/sourcegraph/internal/authz/subrepoperms"
@@ -41,27 +40,6 @@ func Init(
 	database.AuthzWith = func(other basestore.ShareableStore) database.AuthzStore {
 		return database.NewAuthzStore(observationCtx.Logger, db, clock)
 	}
-
-	// TODO(nsc): use c
-	// Report any authz provider problems in external configs.
-	conf.ContributeWarning(func(cfg conftypes.SiteConfigQuerier) (problems conf.Problems) {
-		_, providers, seriousProblems, warnings, _ := providers.ProvidersFromConfig(ctx, cfg, db)
-		problems = append(problems, conf.NewExternalServiceProblems(seriousProblems...)...)
-
-		// Validating the connection may make a cross service call, so we should use an
-		// internal actor.
-		ctx := actor.WithInternalActor(ctx)
-
-		// Add connection validation issue
-		for _, p := range providers {
-			if err := p.ValidateConnection(ctx); err != nil {
-				warnings = append(warnings, fmt.Sprintf("%s provider %q: %s", p.ServiceType(), p.ServiceID(), err))
-			}
-		}
-
-		problems = append(problems, conf.NewExternalServiceProblems(warnings...)...)
-		return problems
-	})
 
 	enterpriseServices.PermissionsGitHubWebhook = webhooks.NewGitHubWebhook(log.Scoped("PermissionsGitHubWebhook"))
