+++ cmd/frontend/internal/batches/BUILD.bazel
@@ -11,9 +11,7 @@ go_library(
         "//cmd/frontend/internal/batches/resolvers",
         "//cmd/frontend/internal/batches/webhooks",
         "//internal/batches/store",
-        "//internal/batches/types/scheduler/window",
         "//internal/codeintel",
-        "//internal/conf",
         "//internal/conf/conftypes",
         "//internal/database",
         "//internal/encryption/keyring",
