+++ cmd/frontend/internal/batches/init.go
@@ -10,9 +10,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/batches/resolvers"
 	"github.com/sourcegraph/sourcegraph/cmd/frontend/internal/batches/webhooks"
 	"github.com/sourcegraph/sourcegraph/internal/batches/store"
-	"github.com/sourcegraph/sourcegraph/internal/batches/types/scheduler/window"
 	"github.com/sourcegraph/sourcegraph/internal/codeintel"
-	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/encryption/keyring"
@@ -31,15 +29,6 @@ func Init(
 	_ conftypes.UnifiedWatchable,
 	enterpriseServices *enterprise.Services,
 ) error {
-	// Validate site configuration.
-	conf.ContributeValidator(func(c conftypes.SiteConfigQuerier) (problems conf.Problems) {
-		if _, err := window.NewConfiguration(c.SiteConfig().BatchChangesRolloutWindows); err != nil {
-			problems = append(problems, conf.NewSiteProblem(err.Error()))
-		}
-
-		return
-	})
-
 	// Initialize store.
 	bstore := store.New(db, observationCtx, keyring.Default().BatchChangesCredentialKey)
