+++ internal/conf/reposource/BUILD.bazel
@@ -29,7 +29,6 @@ go_library(
     deps = [
         "//internal/api",
         "//internal/conf",
-        "//internal/conf/conftypes",
         "//internal/extsvc",
         "//internal/lazyregexp",
         "//lib/errors",
