+++ internal/conf/validation/BUILD.bazel
@@ -1,0 +1,59 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library")
+load("//dev:go_defs.bzl", "go_test")
+
+go_library(
+    name = "validation",
+    srcs = [
+        "app.go",
+        "auth.go",
+        "batches.go",
+        "cody.go",
+        "encryption.go",
+        "highlight.go",
+        "licensing.go",
+        "prometheus.go",
+        "reposource.go",
+        "session.go",
+        "symbols.go",
+        "trace.go",
+        "txemail.go",
+        "validation.go",
+    ],
+    importpath = "github.com/sourcegraph/sourcegraph/internal/conf/validation",
+    visibility = ["//:__subpackages__"],
+    deps = [
+        "//internal/actor",
+        "//internal/authz/providers",
+        "//internal/batches/types/scheduler/window",
+        "//internal/conf",
+        "//internal/conf/conftypes",
+        "//internal/conf/deploy",
+        "//internal/ctags_config",
+        "//internal/database",
+        "//internal/encryption/keyring",
+        "//internal/highlight",
+        "//internal/licensing",
+        "//internal/src-prometheus",
+        "//internal/txemail",
+        "//lib/codeintel/languages",
+        "//lib/errors",
+        "//schema",
+        "@com_github_grafana_regexp//:regexp",
+    ],
+)
+
+go_test(
+    name = "validation_test",
+    srcs = [
+        "auth_test.go",
+        "prometheus_test.go",
+        "txemail_test.go",
+    ],
+    embed = [":validation"],
+    deps = [
+        "//internal/conf",
+        "//internal/src-prometheus",
+        "//schema",
+        "@com_github_hexops_autogold_v2//:autogold",
+    ],
+)
