+++ internal/conf/validation/auth_test.go
@@ -1,4 +1,4 @@
-package auth
+package validation
 
 import (
 	"testing"
@@ -7,7 +7,7 @@ import (
 	"github.com/sourcegraph/sourcegraph/schema"
 )
 
-func TestValidateCustom(t *testing.T) {
+func TestValidateAuthProvidersConfig(t *testing.T) {
 	tests := map[string]struct {
 		input        conf.Unified
 		wantProblems conf.Problems
@@ -40,7 +40,37 @@ func TestValidateCustom(t *testing.T) {
 	}
 	for name, test := range tests {
 		t.Run(name, func(t *testing.T) {
-			conf.TestValidator(t, test.input, validateConfig, test.wantProblems)
+			conf.TestValidator(t, test.input, validateAuthProvidersConfig, test.wantProblems)
+		})
+	}
+}
+
+func TestValidateBuiltinAuthConfig(t *testing.T) {
+	tests := map[string]struct {
+		input        conf.Unified
+		wantProblems conf.Problems
+	}{
+		"single": {
+			input: conf.Unified{SiteConfiguration: schema.SiteConfiguration{
+				AuthProviders: []schema.AuthProviders{
+					{Builtin: &schema.BuiltinAuthProvider{Type: "builtin"}},
+				},
+			}},
+			wantProblems: nil,
+		},
+		"multiple": {
+			input: conf.Unified{SiteConfiguration: schema.SiteConfiguration{
+				AuthProviders: []schema.AuthProviders{
+					{Builtin: &schema.BuiltinAuthProvider{Type: "builtin"}},
+					{Builtin: &schema.BuiltinAuthProvider{Type: "builtin"}},
+				},
+			}},
+			wantProblems: conf.NewSiteProblems("at most 1"),
+		},
+	}
+	for name, test := range tests {
+		t.Run(name, func(t *testing.T) {
+			conf.TestValidator(t, test.input, validateBuiltinAuthConfig, test.wantProblems)
 		})
 	}
 }
