+++ internal/conf/validation/batches.go
@@ -1,0 +1,18 @@
+package validation
+
+import (
+	"github.com/sourcegraph/sourcegraph/internal/batches/types/scheduler/window"
+	"github.com/sourcegraph/sourcegraph/internal/conf"
+	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
+)
+
+func init() {
+	// Validate site configuration.
+	conf.ContributeValidator(func(c conftypes.SiteConfigQuerier) (problems conf.Problems) {
+		if _, err := window.NewConfiguration(c.SiteConfig().BatchChangesRolloutWindows); err != nil {
+			problems = append(problems, conf.NewSiteProblem(err.Error()))
+		}
+
+		return
+	})
+}
