+++ internal/encryption/keyring/ring.go
@@ -2,11 +2,9 @@ package keyring
 
 import (
 	"context"
-	"fmt"
 	"sync"
 
 	"github.com/sourcegraph/sourcegraph/internal/conf"
-	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 	"github.com/sourcegraph/sourcegraph/internal/encryption"
 	"github.com/sourcegraph/sourcegraph/internal/encryption/awskms"
 	"github.com/sourcegraph/sourcegraph/internal/encryption/cache"
@@ -50,13 +48,6 @@ func Init(ctx context.Context) error {
 		defaultRing = *ring
 		mu.Unlock()
 	}
-
-	conf.ContributeValidator(func(cfg conftypes.SiteConfigQuerier) conf.Problems {
-		if _, err := NewRing(ctx, cfg.SiteConfig().EncryptionKeys); err != nil {
-			return conf.Problems{conf.NewSiteProblem(fmt.Sprintf("Invalid encryption.keys config: %s", err))}
-		}
-		return nil
-	})
 
 	conf.Watch(func() {
 		newConfig := conf.Get().EncryptionKeys
