+++ internal/session/BUILD.bazel
@@ -12,7 +12,6 @@ go_library(
     deps = [
         "//internal/actor",
         "//internal/conf",
-        "//internal/conf/conftypes",
         "//internal/database",
         "//internal/env",
         "//internal/errcode",
