+++ internal/tracer/BUILD.bazel
@@ -15,7 +15,6 @@ go_library(
     importpath = "github.com/sourcegraph/sourcegraph/internal/tracer",
     visibility = ["//:__subpackages__"],
     deps = [
-        "//internal/conf",
         "//internal/conf/conftypes",
         "//internal/env",
         "//internal/hostname",
