+++ internal/tracer/tracer.go
@@ -1,9 +1,7 @@
 package tracer
 
 import (
-	"fmt"
 	"sync/atomic"
-	"text/template"
 
 	"github.com/go-logr/logr"
 	"github.com/go-logr/stdr"
@@ -12,7 +10,6 @@ import (
 	"go.opentelemetry.io/otel"
 	"go.uber.org/automaxprocs/maxprocs"
 
-	"github.com/sourcegraph/sourcegraph/internal/conf"
 	"github.com/sourcegraph/sourcegraph/internal/conf/conftypes"
 	"github.com/sourcegraph/sourcegraph/internal/env"
 	"github.com/sourcegraph/sourcegraph/internal/hostname"
@@ -125,18 +122,6 @@ func Init(logger log.Logger, c WatchableConfigurationSource) {
 	// Initially everything is disabled since we haven't read conf yet - start a goroutine
 	// that watches for updates to configure the undelrying provider and debugMode.
 	go c.Watch(newConfWatcher(logger, c, provider, newOtelSpanProcessor, debugMode))
-
-	// Contribute validation for tracing package
-	conf.ContributeWarning(func(c conftypes.SiteConfigQuerier) conf.Problems {
-		tracing := c.SiteConfig().ObservabilityTracing
-		if tracing == nil || tracing.UrlTemplate == "" {
-			return nil
-		}
-		if _, err := template.New("").Parse(tracing.UrlTemplate); err != nil {
-			return conf.NewSiteProblems(fmt.Sprintf("observability.tracing.traceURL is not a valid template: %s", err.Error()))
-		}
-		return nil
-	})
 }
 
 type toggledLogrSink struct {
