+++ internal/txemail/BUILD.bazel
@@ -12,7 +12,6 @@ go_library(
     visibility = ["//:__subpackages__"],
     deps = [
         "//internal/conf",
-        "//internal/conf/conftypes",
         "//internal/txemail/txtypes",
         "//lib/errors",
         "//schema",
@@ -27,14 +26,12 @@ go_test(
     name = "txemail_test",
     timeout = "short",
     srcs = [
-        "siteconfig_test.go",
         "template_test.go",
         "txemail_test.go",
     ],
     embed = [":txemail"],
     deps = [
         "//internal/txemail/txtypes",
-        "//schema",
         "@com_github_google_go_cmp//cmp",
         "@com_github_hexops_autogold_v2//:autogold",
         "@com_github_jordan_wright_email//:email",
